Moved POEApi.Model.Tests to Tests subdirectory.
