+++ Tests/POEApi.Model.Tests/Builders/JSONProxyItemBuilder.cs
