+++ Tests/POEApi.Model.Tests/GearTests.cs
