+++ Tests/POEApi.Model.Tests/IFilterTests.cs
