+++ Tests/POEApi.Model.Tests/ItemFactoryTests.cs
