+++ Tests/POEApi.Model.Tests/TestData/SampleCharacter.json
