+++ Tests/POEApi.Model.Tests/TestData/SampleCharacterExpiredName.json
