+++ Tests/POEApi.Model.Tests/TestData/SampleCurrencyTabWithShards.json
