+++ Tests/POEApi.Model.Tests/TestData/SampleInventory.json
