+++ Tests/POEApi.Model.Tests/TestData/SampleStashWithEssences.json
