+++ Tests/POEApi.Model.Tests/TestData/SampleStashWithLitheBlade.json
