+++ Tests/POEApi.Model.Tests/TestData/SampleStashWithRelic.json
