+++ Tests/POEApi.Model.Tests/TestData/SampleStashWithRemnantOfCorruption.json
