+++ Tests/POEApi.Model.Tests/TestData/SampleStashWithSaintlyChainmail.json
