+++ Tests/POEApi.Model.Tests/UnitTestHelper.cs
