+++ Tests/POEApi.Model.Tests/packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Castle.Core" version="3.3.3" targetFramework="net45" />
+  <package id="FluentAssertions" version="4.19.4" targetFramework="net45" />
+  <package id="Moq" version="4.5.22" targetFramework="net45" />
+</packages>
